Progress on making ObjectDB thread-safe. Previously, only one thread could work with the object database at a time. Mostly fixed now.
TODO: allow more than one thread to create object groups at a time.

+++ src/system/fabric/common/util/MutableLong.java
@@ -1,13 +1,13 @@
 package fabric.common.util;
 
-public class MutableInteger {
-  public int value;
+public class MutableLong {
+  public long value;
 
-  public MutableInteger() {
+  public MutableLong() {
     this(0);
   }
 
-  public MutableInteger(int value) {
+  public MutableLong(long value) {
     this.value = value;
   }
